test: Move util functions over into utils.js
